backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
scikit-learn
python-dotenv
# Add your Khmer NLP library here (e.g., khmer-nginx or similar)